SOF-1700 Move HttpActionTriggerService to correct folder
